+++ rna_tools/rna_tools_lib.py
@@ -970,6 +970,7 @@ class RNAStructure:
     def get_rnapuzzle_ready(self, renumber_residues=True, fix_missing_atoms=True,
                             rename_chains=True,
                             report_missing_atoms=True, keep_hetatm=False, backbone_only=False,
+                            no_backbone = False, bases_only = False,
                             verbose=False):  # :, ready_for="RNAPuzzle"):
         """Get rnapuzzle (SimRNA) ready structure.
 
@@ -1022,6 +1023,16 @@ class RNAStructure:
             A_ATOMS = "P OP1 OP2 O5' C5' C4' O4' C3' O3' C2' O2' C1'".split()
             U_ATOMS = "P OP1 OP2 O5' C5' C4' O4' C3' O3' C2' O2' C1'".split()
             C_ATOMS = "P OP1 OP2 O5' C5' C4' O4' C3' O3' C2' O2' C1'".split()
+        elif no_backbone:
+            G_ATOMS = "C5' C4' O4' C3' O3' C2' O2' C1' N9 C8 N7 C5 C6 O6 N1 C2 N2 N3 C4".split()
+            A_ATOMS = "C5' C4' O4' C3' O3' C2' O2' C1' N9 C8 N7 C5 C6 N6 N1 C2 N3 C4".split()
+            U_ATOMS = "C5' C4' O4' C3' O3' C2' O2' C1' N1 C2 O2 N3 C4 O4 C5 C6".split()
+            C_ATOMS = "C5' C4' O4' C3' O3' C2' O2' C1' N1 C2 O2 N3 C4 N4 C5 C6".split()
+        elif bases_only:
+            G_ATOMS = "N9 C8 N7 C5 C6 O6 N1 C2 N2 N3 C4".split()
+            A_ATOMS = "N9 C8 N7 C5 C6 N6 N1 C2 N3 C4".split()
+            U_ATOMS = "N1 C2 O2 N3 C4 O4 C5 C6".split()
+            C_ATOMS = "N1 C2 O2 N3 C4 N4 C5 C6".split()
         else:
             G_ATOMS = "P OP1 OP2 O5' C5' C4' O4' C3' O3' C2' O2' C1' N9 C8 N7 C5 C6 O6 N1 C2 N2 N3 C4".split()
             A_ATOMS = "P OP1 OP2 O5' C5' C4' O4' C3' O3' C2' O2' C1' N9 C8 N7 C5 C6 N6 N1 C2 N3 C4".split()