+++ rna_tools/tools/mq/ClashScore/test.sh
@@ -1,0 +1,2 @@
+python ClashScore.py | tee ../../../output/ClashScore.txt
+./rna_clashscore.py ../test/1xjr*.pdb | tee ../../../output/run_clashscore.txt