+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - RNA3DCNN
         - QRNA
         - FARNA
-        - AnalyzeGeometry and ClashScore
+        - AnalyzeGeometry and ClashScore (add rna_clashscore.py app)
         - eSCORE (barnaba)
         - 3dRNAscore
         - libs
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     Changed:
     
         rna_tools_lib.py (RNAStructure) when rpr loads structure with biopython the error is not suppressed
+        
+
 
 # 3.6